[#5] impl: find_interface_and_append_method
