+++ src/handle/boilerplates.rs
@@ -1,4 +1,4 @@
-use crate::config::RootConfig;
+use crate::{config::RootConfig, utils::split_to_pair};
 use convert_case::{Case, Casing};
 
 pub fn generate_usecase_interface_type(domain: String, config_file: &RootConfig) -> String {
@@ -42,4 +42,39 @@ pub fn generate_usecase_interface_method(
     };
 
     method_signature
+}
+
+pub fn split_to_pair(s: &str, pattern: &str) -> Option<(String, String)> {
+    let mut split = s.split(pattern);
+
+    let first = split.next()?.to_owned();
+
+    let mut second = split.next()?.to_owned();
+
+    while let Some(s) = split.next() {
+        second = second.to_string() + s;
+    }
+
+    Some((first, second))
+}
+
+pub fn find_interface_and_append_method(
+    code: &str,
+    interface_typename: &str,
+    method_code: &str,
+) -> String {
+    let typename_splited_pair = split_to_pair(code, &interface_typename).unwrap();
+
+    let mut new_content = String::new();
+
+    new_content.push_str(typename_splited_pair.0.as_str());
+    new_content.push_str(&interface_typename);
+
+    let brace_splited_pair = split_to_pair(&typename_splited_pair.1, "}").unwrap();
+    new_content.push_str(brace_splited_pair.0.as_str());
+    new_content.push_str(format!("    {}\n", method_code).as_str());
+    new_content.push_str("}");
+    new_content.push_str(brace_splited_pair.1.as_str());
+
+    new_content
 }
